libnss_tcb: Initialize or rewind dirstream from inside setspent(3).
On first call to setspent(3) initialize the directory stream properly;
on subsequent calls use rewinddir(3) to reset the position of the directory
stream to the beginning of the directory, and also update the existing
directory stream to refer to the current state of the underlying directory
it operates on.  As all internal functions are operating on thread-local
storage now, this operation will be safe, since it will emit no effects
outside of the thread calling the setspent(3) function itself.

Signed-off-by: Björn Esser <besser82@fedoraproject.org>